#26 refactor : feature:main 으로 분리
