+++ feature/main/src/main/kotlin/com/titi/feature/main/Screen.kt
@@ -1,4 +1,4 @@
-package com.titi.navigation
+package com.titi.feature.main
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
