+++ feature/main/src/main/kotlin/com/titi/feature/main/di/ViewModelModule.kt
@@ -1,9 +1,9 @@
-package com.titi.di
+package com.titi.feature.main.di
 
 import com.airbnb.mvrx.hilt.AssistedViewModelFactory
 import com.airbnb.mvrx.hilt.MavericksViewModelComponent
 import com.airbnb.mvrx.hilt.ViewModelKey
-import com.titi.MainViewModel
+import com.titi.feature.main.ui.main.MainViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
