+++ feature/main/src/main/kotlin/com/titi/feature/main/model/SplashResultState.kt
@@ -1,4 +1,4 @@
-package com.titi
+package com.titi.feature.main.model
 
 import com.titi.domain.color.model.TimeColor
 import com.titi.domain.time.model.RecordTimes
