+++ feature/main/src/main/kotlin/com/titi/feature/main/ui/main/MainViewModel.kt
@@ -1,25 +1,17 @@
-package com.titi
+package com.titi.feature.main.ui.main
 
 import android.util.Log
-import com.airbnb.mvrx.MavericksState
 import com.airbnb.mvrx.MavericksViewModel
 import com.airbnb.mvrx.MavericksViewModelFactory
 import com.airbnb.mvrx.hilt.AssistedViewModelFactory
 import com.airbnb.mvrx.hilt.hiltMavericksViewModelFactory
-import com.titi.domain.color.model.TimeColor
 import com.titi.domain.color.usecase.GetTimeColorFlowUseCase
-import com.titi.domain.time.model.RecordTimes
 import com.titi.domain.time.usecase.GetRecordTimesFlowUseCase
+import com.titi.feature.main.model.MainUiState
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.flow.catch
-
-data class MainUiState(
-    val recordTimes: RecordTimes = RecordTimes(),
-    val timeColor: TimeColor = TimeColor(),
-    val bottomNavigationPosition: Int = 0,
-) : MavericksState
 
 class MainViewModel @AssistedInject constructor(
     @Assisted initialState: MainUiState,
