+++ feature/main/src/main/kotlin/com/titi/feature/main/ui/splash/SplashScreen.kt
@@ -1,10 +1,11 @@
-package com.titi
+package com.titi.feature.main.ui.splash
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.titi.feature.main.model.SplashResultState
 
 @Composable
 fun SplashScreen(
