+++ feature/main/src/main/res/values/themes.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <style name="Theme.TiTi" parent="android:Theme.Material.Light.NoActionBar" />
 </resources>
